# main-api/artifacts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Doc",
    "description": "Documentacao da API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {}
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate within the Corporation resolved from the trusted host",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 300
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024,
                    "writeOnly": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "message", "data"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "required": ["accessToken", "refreshToken", "expiresIn"],
                      "properties": {
                        "accessToken": { "type": "string" },
                        "refreshToken": { "type": "string" },
                        "expiresIn": { "type": "integer", "enum": [900] }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "summary": "Rotate the current browser refresh credential",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 32,
                    "writeOnly": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "message", "data"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "required": ["accessToken", "refreshToken", "expiresIn"],
                      "properties": {
                        "accessToken": { "type": "string" },
                        "refreshToken": { "type": "string" },
                        "expiresIn": { "type": "integer", "enum": [900] }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "Revoke the current persisted Session",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "message", "data"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "required": ["revoked"],
                      "properties": {
                        "revoked": { "type": "boolean", "enum": [true] }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "summary": "Inspect the authenticated persisted Session",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "message", "data"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "required": [
                        "userId",
                        "corporationId",
                        "sessionId",
                        "role",
                        "companyId",
                        "scope",
                        "selectedCompany"
                      ],
                      "properties": {
                        "userId": { "type": "string", "format": "uuid" },
                        "corporationId": { "type": "string", "format": "uuid" },
                        "sessionId": { "type": "string", "format": "uuid" },
                        "role": { "type": "string", "enum": ["MASTER_ADMIN"] },
                        "companyId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "scope": {
                          "type": "string",
                          "enum": ["corporation-scoped", "company-scoped"]
                        },
                        "selectedCompany": {
                          "type": "object",
                          "nullable": true,
                          "required": ["id", "name"],
                          "properties": {
                            "id": { "type": "string", "format": "uuid" },
                            "name": { "type": "string" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/companies": {
      "get": {
        "summary": "List active Companies for the authenticated Corporation",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "message", "data"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "required": ["companies"],
                      "properties": {
                        "companies": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": ["id", "name"],
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/session/company": {
      "put": {
        "summary": "Select or change the active Company for the current Session",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["companyId"],
                "properties": {
                  "companyId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "message", "data"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "required": [
                        "accessToken",
                        "expiresIn",
                        "selectedCompany"
                      ],
                      "properties": {
                        "accessToken": { "type": "string" },
                        "expiresIn": { "type": "integer", "enum": [900] },
                        "selectedCompany": {
                          "type": "object",
                          "required": ["id", "name"],
                          "properties": {
                            "id": { "type": "string", "format": "uuid" },
                            "name": { "type": "string" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "code", "message", "requestId"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [false] },
                    "code": { "type": "string" },
                    "message": { "type": "string" },
                    "details": { "type": "object", "nullable": true },
                    "requestId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [{ "url": "http://0.0.0.0:3333", "description": "Servidor local" }]
}